requirements.txt
pytest
pytest-cov
pyyaml
python-dotenv==1.1.0
slack_bolt==1.23.0
slack_sdk==3.35.0
PyYAML
